+++ src/zabbix_server/lld/lld_trigger.c
@@ -2653,7 +2653,7 @@ static int	lld_triggers_save(zbx_uint64_t hostid, const zbx_vector_ptr_t *trigge
 		{
 			zbx_audit_trigger_create_entry(AUDIT_ACTION_UPDATE, trigger->triggerid,
 					(NULL == trigger->description_orig) ? trigger->description :
-					trigger->description_orig, (int)ZBX_FLAG_DISCOVERY_CREATED);
+					trigger->description_orig, ZBX_FLAG_DISCOVERY_CREATED);
 		}
 
 		for (j = 0; j < trigger->functions.values_num; j++)
@@ -2854,15 +2854,18 @@ static int	lld_triggers_save(zbx_uint64_t hostid, const zbx_vector_ptr_t *trigge
 					trigger->opdata, trigger->event_name);
 
 			zbx_audit_trigger_create_entry(AUDIT_ACTION_ADD,triggerid, trigger->description,
-					(int)ZBX_FLAG_DISCOVERY_CREATED);
+					ZBX_FLAG_DISCOVERY_CREATED);
+
 			zbx_audit_trigger_update_json_add_data(triggerid, 0, (int)trigger_prototype->recovery_mode,
 					trigger->status, trigger_prototype->type, TRIGGER_VALUE_OK,
 					TRIGGER_STATE_NORMAL, trigger->priority, trigger->comments, trigger->url,
 					ZBX_FLAG_DISCOVERY_CREATED, trigger_prototype->correlation_mode,
 					trigger->correlation_tag, trigger_prototype->manual_close, trigger->opdata, 0,
 					trigger->event_name);
+
 			zbx_audit_trigger_update_json_add_expr(triggerid,
 					(int)ZBX_FLAG_DISCOVERY_CREATED, trigger->expression);
+
 			zbx_audit_trigger_update_json_add_rexpr(triggerid,
 					(int)ZBX_FLAG_DISCOVERY_CREATED, trigger->recovery_expression);
